+++ cache-node/src/main/kotlin/cache/distributed/IDistributedCache.kt
@@ -2,8 +2,9 @@ package cache.distributed
 
 import KeyVersionPair
 import NodeId
-import receiver.MemoryUsageInfo
-import receiver.SystemInfo
+import cache.local.CacheInfo
+import receiver.ReceiverUsageInfo
+import sender.SenderUsageInfo
 
 /**
  * An interface specifying the behavior of a distributed data cache.
@@ -40,4 +41,19 @@ interface IDistributedCache {
         val usage = allocatedMemory/(maxMemory * 1.0)
         return MemoryUsageInfo(allocatedMemory, maxMemory, usage)
     }
+
+    /**
+     * Client response giving memory usage of the JVM.
+     */
+    data class MemoryUsageInfo(val allocated: Long, val max: Long, val usage: Double)
+
+    /**
+     * Encapsulates information about the usage of this node into one object
+     */
+    data class SystemInfo(
+        val nodeId: Int,
+        val memUsage: MemoryUsageInfo,
+        val cacheInfo: CacheInfo,
+        var receiverUsageInfo: ReceiverUsageInfo,
+        val senderUsageInfo: SenderUsageInfo?)
 }
